+++ database/migrations/2026_03_21_060525_create_order_items_table.php
@@ -8,16 +8,18 @@ return new class extends \Illuminate\Database\Migrations\Migration
 {
     public function up()
     {
-        Schema::create('order_items', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('order_id')->constrained()->onDelete('cascade');
-            $table->foreignId('item_id')->constrained()->onDelete('cascade');
-            $table->foreignId('item_variant_id')->nullable()->constrained()->onDelete('set null');
-            $table->decimal('price', 10, 2);
-            $table->integer('quantity');
-            $table->decimal('total', 10, 2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('order_items')) {
+            Schema::create('order_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('order_id')->constrained()->onDelete('cascade');
+                $table->foreignId('item_id')->constrained()->onDelete('cascade');
+                $table->foreignId('item_variant_id')->nullable()->constrained()->onDelete('set null');
+                $table->decimal('price', 10, 2);
+                $table->integer('quantity');
+                $table->decimal('total', 10, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
